requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
psycopg2-binary==2.9.10
python-dotenv==1.0.0
Jinja2==3.1.6
gunicorn==23.0.0
